// frontend/src/metabase/query_builder/components/Canvas/types.ts
export type ProgrammingLanguageOptions =
  | "typescript"
  | "javascript"
  | "cpp"
  | "java"
  | "php"
  | "python"
  | "html"
  | "sql"
  | "other";
// ...
export interface ArtifactCodeV3 {
  index: number;
  type: "code";
  title: string;
  description?: string;
  language: ProgrammingLanguageOptions;
  code: string;
}